Actually return a sign-up URL
Currently the sign-up e-mail just sends a token.  The frontend doesn't
give us a callback URL for reset like it does for sign-up.  This is
a little bit hacky to look at the referer, but it works and we'll
generate the correct URL based on a Referer sent from the browser.

This gives a user something to click.